Add explicit cast to fix unsigned truncation warning
Cast y to uint32_t in the remainder adjustment `r -= (uint32_t)y`
to match the type of r, suppressing the implicit truncation warning
from MSVC (C4267).


Mirrored-from: 9e9f289d20017f213df4854aca86a0f3c282ba9b

+++ src/include/nccl_device/utility.h
@@ -256,7 +256,7 @@ NCCL_HOST_DEVICE_INLINE void idivmodFast32(uint32_t *quo, uint32_t *rem, uint32_
 NCCL_HOST_DEVICE_INLINE void idivmodFast64(uint64_t *quo, uint64_t *rem, uint64_t x, uint64_t y, uint64_t yrcp) {
   uint32_t q = (uint32_t)(yrcp == 0 ? x : mul64hi(x, yrcp));
   uint32_t r = (uint32_t)(x - y*q);
-  if (r >= y) { q += 1; r -= y; }
+  if (r >= y) { q += 1; r -= (uint32_t)y; }
   *quo = q;
   *rem = r;
 }
